fix(CustomEndpoints): more major refactors, code splitting
